prevented users from reserving more than 1 bag at a time. cleaned up thunks
