+++ src/components/Cart/cartBagSlice.js
@@ -49,44 +49,7 @@ export const placeBagInCartAsync = createAsyncThunk(
   }
 );
 
-export const fetchOrderByStatusAsync = createAsyncThunk(
-  "cart",
-  async (userId, status) => {
-    try {
-      //fetch based on userId and "shopping" status
-      const ordersRef = collection(db, "orders");
 
-      const q = query(
-        ordersRef,
-        where("userId", "==", userId),
-        where("status", "==", "shopping")
-      );
-
-      const querySnapshot = await getDocs(q);
-      const orders = [];
-      querySnapshot.forEach((doc) => {
-        orders.push({...doc.data(), id: doc.id})
-      });
-
-      return orders;
-    } catch (err) {
-      console.error(err);
-    }
-  }
-);
-
-// delete bag from cart
-export const deleteOrderAsync = createAsyncThunk(
-  "deleteOrder",
-  async (orderId) => {
-    try {
-      const orderRef = doc(db, "orders", orderId);
-      await deleteDoc(orderRef);
-    } catch (err) {
-      console.error(err);
-    }
-  }
-);
 
 const initialState = [];
 export const cartBagSlice = createSlice({
@@ -97,15 +60,7 @@ export const cartBagSlice = createSlice({
     builder
       .addCase(placeBagInCartAsync.fulfilled, (state, action) => {
         return action.payload;
-      })
-      .addCase(fetchOrderByStatusAsync.fulfilled, (state, action) => {
-        return action.payload;
-      })
-      .addCase(deleteOrderAsync.fulfilled, (state, action) => {
-        const orderId = action.payload;
-
-        return state.filter((order) => order.id !== orderId);
-      });
+      })  
   },
 });
